fix : Remove obsolete migration files.
