--- alembic/versions/05cfcb09c1ff_feat_add_config.py
@@ -1,197 +1,0 @@
-"""feat:Add config
-
-Revision ID: 05cfcb09c1ff
-Revises: 
-Create Date: 2025-10-28 01:37:51.548562
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '05cfcb09c1ff'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('companies',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('company_number', sa.Integer(), nullable=False),
-    sa.Column('company_name', sa.String(), nullable=False),
-    sa.Column('managers', postgresql.JSONB(astext_type=sa.Text()), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_companies_company_name'), 'companies', ['company_name'], unique=False)
-    op.create_index(op.f('ix_companies_company_number'), 'companies', ['company_number'], unique=True)
-    op.create_index(op.f('ix_companies_id'), 'companies', ['id'], unique=False)
-    op.create_table('countries',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('code', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_countries_code'), 'countries', ['code'], unique=False)
-    op.create_index(op.f('ix_countries_id'), 'countries', ['id'], unique=False)
-    op.create_index(op.f('ix_countries_name'), 'countries', ['name'], unique=False)
-    op.create_table('positions',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('lv1_name', sa.String(), nullable=False),
-    sa.Column('lv2_name', sa.String(), nullable=False),
-    sa.Column('lv3_name', sa.String(), nullable=False),
-    sa.Column('lv4_name', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_positions_id'), 'positions', ['id'], unique=False)
-    op.create_index(op.f('ix_positions_lv1_name'), 'positions', ['lv1_name'], unique=False)
-    op.create_index(op.f('ix_positions_lv2_name'), 'positions', ['lv2_name'], unique=False)
-    op.create_index(op.f('ix_positions_lv3_name'), 'positions', ['lv3_name'], unique=False)
-    op.create_index(op.f('ix_positions_lv4_name'), 'positions', ['lv4_name'], unique=False)
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('passport_certi', sa.Boolean(), nullable=False),
-    sa.Column('company_info', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_index(op.f('ix_users_passport_certi'), 'users', ['passport_certi'], unique=False)
-    op.create_table('profiles',
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('profile_image_url', sa.String(), nullable=True),
-    sa.Column('location', sa.String(), nullable=True),
-    sa.Column('introduction', sa.String(), nullable=True),
-    sa.Column('address', sa.String(), nullable=True),
-    sa.Column('position_id', sa.Integer(), nullable=True),
-    sa.Column('job_status', sa.String(), nullable=True),
-    sa.Column('portfolio_url', sa.String(), nullable=True),
-    sa.Column('birth_date', sa.Date(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('country_id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['country_id'], ['countries.id'], ),
-    sa.ForeignKeyConstraint(['position_id'], ['positions.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('user_id')
-    )
-    op.create_index(op.f('ix_profiles_address'), 'profiles', ['address'], unique=False)
-    op.create_index(op.f('ix_profiles_birth_date'), 'profiles', ['birth_date'], unique=False)
-    op.create_index(op.f('ix_profiles_country_id'), 'profiles', ['country_id'], unique=False)
-    op.create_index(op.f('ix_profiles_introduction'), 'profiles', ['introduction'], unique=False)
-    op.create_index(op.f('ix_profiles_job_status'), 'profiles', ['job_status'], unique=False)
-    op.create_index(op.f('ix_profiles_location'), 'profiles', ['location'], unique=False)
-    op.create_index(op.f('ix_profiles_name'), 'profiles', ['name'], unique=False)
-    op.create_index(op.f('ix_profiles_portfolio_url'), 'profiles', ['portfolio_url'], unique=False)
-    op.create_index(op.f('ix_profiles_position_id'), 'profiles', ['position_id'], unique=False)
-    op.create_index(op.f('ix_profiles_profile_image_url'), 'profiles', ['profile_image_url'], unique=False)
-    op.create_index(op.f('ix_profiles_user_id'), 'profiles', ['user_id'], unique=False)
-    op.create_table('refresh_tokens',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('token', sa.String(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('expires_at', sa.DateTime(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_refresh_tokens_expires_at'), 'refresh_tokens', ['expires_at'], unique=False)
-    op.create_index(op.f('ix_refresh_tokens_id'), 'refresh_tokens', ['id'], unique=False)
-    op.create_index(op.f('ix_refresh_tokens_token'), 'refresh_tokens', ['token'], unique=True)
-    op.create_index(op.f('ix_refresh_tokens_user_id'), 'refresh_tokens', ['user_id'], unique=False)
-    op.create_table('account_config',
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('sns_message_notice', sa.Boolean(), nullable=False),
-    sa.Column('email_notice', sa.Boolean(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['profiles.user_id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('user_id')
-    )
-    op.create_index(op.f('ix_account_config_email_notice'), 'account_config', ['email_notice'], unique=False)
-    op.create_index(op.f('ix_account_config_sns_message_notice'), 'account_config', ['sns_message_notice'], unique=False)
-    op.create_index(op.f('ix_account_config_user_id'), 'account_config', ['user_id'], unique=False)
-    op.create_table('contacts',
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('phone_number', sa.String(), nullable=True),
-    sa.Column('github_url', sa.String(), nullable=True),
-    sa.Column('linkedin_url', sa.String(), nullable=True),
-    sa.Column('website_url', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['profiles.user_id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('user_id')
-    )
-    op.create_index(op.f('ix_contacts_github_url'), 'contacts', ['github_url'], unique=False)
-    op.create_index(op.f('ix_contacts_linkedin_url'), 'contacts', ['linkedin_url'], unique=False)
-    op.create_index(op.f('ix_contacts_phone_number'), 'contacts', ['phone_number'], unique=False)
-    op.create_index(op.f('ix_contacts_user_id'), 'contacts', ['user_id'], unique=False)
-    op.create_index(op.f('ix_contacts_website_url'), 'contacts', ['website_url'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_contacts_website_url'), table_name='contacts')
-    op.drop_index(op.f('ix_contacts_user_id'), table_name='contacts')
-    op.drop_index(op.f('ix_contacts_phone_number'), table_name='contacts')
-    op.drop_index(op.f('ix_contacts_linkedin_url'), table_name='contacts')
-    op.drop_index(op.f('ix_contacts_github_url'), table_name='contacts')
-    op.drop_table('contacts')
-    op.drop_index(op.f('ix_account_config_user_id'), table_name='account_config')
-    op.drop_index(op.f('ix_account_config_sns_message_notice'), table_name='account_config')
-    op.drop_index(op.f('ix_account_config_email_notice'), table_name='account_config')
-    op.drop_table('account_config')
-    op.drop_index(op.f('ix_refresh_tokens_user_id'), table_name='refresh_tokens')
-    op.drop_index(op.f('ix_refresh_tokens_token'), table_name='refresh_tokens')
-    op.drop_index(op.f('ix_refresh_tokens_id'), table_name='refresh_tokens')
-    op.drop_index(op.f('ix_refresh_tokens_expires_at'), table_name='refresh_tokens')
-    op.drop_table('refresh_tokens')
-    op.drop_index(op.f('ix_profiles_user_id'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_profile_image_url'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_position_id'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_portfolio_url'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_name'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_location'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_job_status'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_introduction'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_country_id'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_birth_date'), table_name='profiles')
-    op.drop_index(op.f('ix_profiles_address'), table_name='profiles')
-    op.drop_table('profiles')
-    op.drop_index(op.f('ix_users_passport_certi'), table_name='users')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_index(op.f('ix_users_email'), table_name='users')
-    op.drop_table('users')
-    op.drop_index(op.f('ix_positions_lv4_name'), table_name='positions')
-    op.drop_index(op.f('ix_positions_lv3_name'), table_name='positions')
-    op.drop_index(op.f('ix_positions_lv2_name'), table_name='positions')
-    op.drop_index(op.f('ix_positions_lv1_name'), table_name='positions')
-    op.drop_index(op.f('ix_positions_id'), table_name='positions')
-    op.drop_table('positions')
-    op.drop_index(op.f('ix_countries_name'), table_name='countries')
-    op.drop_index(op.f('ix_countries_id'), table_name='countries')
-    op.drop_index(op.f('ix_countries_code'), table_name='countries')
-    op.drop_table('countries')
-    op.drop_index(op.f('ix_companies_id'), table_name='companies')
-    op.drop_index(op.f('ix_companies_company_number'), table_name='companies')
-    op.drop_index(op.f('ix_companies_company_name'), table_name='companies')
-    op.drop_table('companies')
-    # ### end Alembic commands ###
