--- alembic/versions/068c79805fce_feat_add_company_posts_table.py
@@ -1,80 +1,0 @@
-"""feat:Add company_posts table
-
-Revision ID: 068c79805fce
-Revises: 1de9b099cc80
-Create Date: 2025-11-11 01:45:55.953382
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '068c79805fce'
-down_revision: Union[str, Sequence[str], None] = '1de9b099cc80'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('company_posts',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('company_id', sa.Integer(), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('content', sa.String(), nullable=False),
-    sa.Column('work_experience', sa.Integer(), nullable=False),
-    sa.Column('position_id', sa.Integer(), nullable=False),
-    sa.Column('education', sa.String(), nullable=False),
-    sa.Column('language', sa.String(), nullable=False),
-    sa.Column('employment_type', sa.String(), nullable=False),
-    sa.Column('work_location', sa.String(), nullable=False),
-    sa.Column('working_hours', sa.Integer(), nullable=False),
-    sa.Column('salary', sa.Integer(), nullable=False),
-    sa.Column('start_date', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('end_date', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['company_id'], ['companies.id'], ),
-    sa.ForeignKeyConstraint(['position_id'], ['positions.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_company_posts_company_id'), 'company_posts', ['company_id'], unique=False)
-    op.create_index(op.f('ix_company_posts_content'), 'company_posts', ['content'], unique=False)
-    op.create_index(op.f('ix_company_posts_education'), 'company_posts', ['education'], unique=False)
-    op.create_index(op.f('ix_company_posts_employment_type'), 'company_posts', ['employment_type'], unique=False)
-    op.create_index(op.f('ix_company_posts_end_date'), 'company_posts', ['end_date'], unique=False)
-    op.create_index(op.f('ix_company_posts_id'), 'company_posts', ['id'], unique=False)
-    op.create_index(op.f('ix_company_posts_language'), 'company_posts', ['language'], unique=False)
-    op.create_index(op.f('ix_company_posts_position_id'), 'company_posts', ['position_id'], unique=False)
-    op.create_index(op.f('ix_company_posts_salary'), 'company_posts', ['salary'], unique=False)
-    op.create_index(op.f('ix_company_posts_start_date'), 'company_posts', ['start_date'], unique=False)
-    op.create_index(op.f('ix_company_posts_title'), 'company_posts', ['title'], unique=False)
-    op.create_index(op.f('ix_company_posts_work_experience'), 'company_posts', ['work_experience'], unique=False)
-    op.create_index(op.f('ix_company_posts_work_location'), 'company_posts', ['work_location'], unique=False)
-    op.create_index(op.f('ix_company_posts_working_hours'), 'company_posts', ['working_hours'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_company_posts_working_hours'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_work_location'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_work_experience'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_title'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_start_date'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_salary'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_position_id'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_language'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_id'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_end_date'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_employment_type'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_education'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_content'), table_name='company_posts')
-    op.drop_index(op.f('ix_company_posts_company_id'), table_name='company_posts')
-    op.drop_table('company_posts')
-    # ### end Alembic commands ###
