--- alembic/versions/1de9b099cc80_feat_add_company_profile_table.py
@@ -1,67 +1,0 @@
-"""feat:Add company_profile table
-
-Revision ID: 1de9b099cc80
-Revises: 4caeeb791f47
-Create Date: 2025-11-08 12:19:32.621695
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '1de9b099cc80'
-down_revision: Union[str, Sequence[str], None] = '4caeeb791f47'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('company_profile',
-    sa.Column('company_id', sa.Integer(), nullable=False),
-    sa.Column('industry_type', sa.String(), nullable=False),
-    sa.Column('employee_count', sa.Integer(), nullable=False),
-    sa.Column('establishment_date', sa.Date(), nullable=False),
-    sa.Column('company_type', sa.String(), nullable=False),
-    sa.Column('insurance', sa.String(), nullable=False),
-    sa.Column('phone_number', sa.BigInteger(), nullable=False),
-    sa.Column('address', sa.String(), nullable=False),
-    sa.Column('website_url', sa.String(), nullable=False),
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['company_id'], ['companies.id'], ),
-    sa.PrimaryKeyConstraint('company_id')
-    )
-    op.create_index(op.f('ix_company_profile_address'), 'company_profile', ['address'], unique=False)
-    op.create_index(op.f('ix_company_profile_company_id'), 'company_profile', ['company_id'], unique=False)
-    op.create_index(op.f('ix_company_profile_company_type'), 'company_profile', ['company_type'], unique=False)
-    op.create_index(op.f('ix_company_profile_email'), 'company_profile', ['email'], unique=False)
-    op.create_index(op.f('ix_company_profile_employee_count'), 'company_profile', ['employee_count'], unique=False)
-    op.create_index(op.f('ix_company_profile_establishment_date'), 'company_profile', ['establishment_date'], unique=False)
-    op.create_index(op.f('ix_company_profile_industry_type'), 'company_profile', ['industry_type'], unique=False)
-    op.create_index(op.f('ix_company_profile_insurance'), 'company_profile', ['insurance'], unique=False)
-    op.create_index(op.f('ix_company_profile_phone_number'), 'company_profile', ['phone_number'], unique=False)
-    op.create_index(op.f('ix_company_profile_website_url'), 'company_profile', ['website_url'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_company_profile_website_url'), table_name='company_profile')
-    op.drop_index(op.f('ix_company_profile_phone_number'), table_name='company_profile')
-    op.drop_index(op.f('ix_company_profile_insurance'), table_name='company_profile')
-    op.drop_index(op.f('ix_company_profile_industry_type'), table_name='company_profile')
-    op.drop_index(op.f('ix_company_profile_establishment_date'), table_name='company_profile')
-    op.drop_index(op.f('ix_company_profile_employee_count'), table_name='company_profile')
-    op.drop_index(op.f('ix_company_profile_email'), table_name='company_profile')
-    op.drop_index(op.f('ix_company_profile_company_type'), table_name='company_profile')
-    op.drop_index(op.f('ix_company_profile_company_id'), table_name='company_profile')
-    op.drop_index(op.f('ix_company_profile_address'), table_name='company_profile')
-    op.drop_table('company_profile')
-    # ### end Alembic commands ###
