+++ alembic/versions/31f887c48484_feat_add_base_table.py
@@ -1,0 +1,288 @@
+"""feat: Add base table
+
+Revision ID: 31f887c48484
+Revises: 
+Create Date: 2025-11-13 23:50:08.298727
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '31f887c48484'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('companies',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('company_number', sa.String(), nullable=False),
+    sa.Column('company_name', sa.String(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_companies_company_name'), 'companies', ['company_name'], unique=False)
+    op.create_index(op.f('ix_companies_company_number'), 'companies', ['company_number'], unique=True)
+    op.create_index(op.f('ix_companies_id'), 'companies', ['id'], unique=False)
+    op.create_table('countries',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('code', sa.String(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_countries_code'), 'countries', ['code'], unique=False)
+    op.create_index(op.f('ix_countries_id'), 'countries', ['id'], unique=False)
+    op.create_index(op.f('ix_countries_name'), 'countries', ['name'], unique=False)
+    op.create_table('positions',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('level', sa.Integer(), nullable=False),
+    sa.Column('parent_id', sa.Integer(), nullable=True),
+    sa.Column('code', sa.String(), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_positions_code'), 'positions', ['code'], unique=False)
+    op.create_index(op.f('ix_positions_id'), 'positions', ['id'], unique=False)
+    op.create_index(op.f('ix_positions_level'), 'positions', ['level'], unique=False)
+    op.create_index(op.f('ix_positions_name'), 'positions', ['name'], unique=False)
+    op.create_index(op.f('ix_positions_parent_id'), 'positions', ['parent_id'], unique=False)
+    op.create_table('users',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('email', sa.String(), nullable=False),
+    sa.Column('passport_certi', sa.Boolean(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
+    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
+    op.create_index(op.f('ix_users_passport_certi'), 'users', ['passport_certi'], unique=False)
+    op.create_table('company_posts',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('company_id', sa.Integer(), nullable=False),
+    sa.Column('title', sa.String(), nullable=False),
+    sa.Column('content', sa.String(), nullable=False),
+    sa.Column('work_experience', sa.Integer(), nullable=False),
+    sa.Column('position_id', sa.Integer(), nullable=False),
+    sa.Column('education', sa.String(), nullable=False),
+    sa.Column('language', sa.String(), nullable=False),
+    sa.Column('employment_type', sa.String(), nullable=False),
+    sa.Column('work_location', sa.String(), nullable=False),
+    sa.Column('working_hours', sa.Integer(), nullable=False),
+    sa.Column('salary', sa.Integer(), nullable=False),
+    sa.Column('start_date', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('end_date', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['company_id'], ['companies.id'], ),
+    sa.ForeignKeyConstraint(['position_id'], ['positions.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_company_posts_company_id'), 'company_posts', ['company_id'], unique=False)
+    op.create_index(op.f('ix_company_posts_content'), 'company_posts', ['content'], unique=False)
+    op.create_index(op.f('ix_company_posts_education'), 'company_posts', ['education'], unique=False)
+    op.create_index(op.f('ix_company_posts_employment_type'), 'company_posts', ['employment_type'], unique=False)
+    op.create_index(op.f('ix_company_posts_end_date'), 'company_posts', ['end_date'], unique=False)
+    op.create_index(op.f('ix_company_posts_id'), 'company_posts', ['id'], unique=False)
+    op.create_index(op.f('ix_company_posts_language'), 'company_posts', ['language'], unique=False)
+    op.create_index(op.f('ix_company_posts_position_id'), 'company_posts', ['position_id'], unique=False)
+    op.create_index(op.f('ix_company_posts_salary'), 'company_posts', ['salary'], unique=False)
+    op.create_index(op.f('ix_company_posts_start_date'), 'company_posts', ['start_date'], unique=False)
+    op.create_index(op.f('ix_company_posts_title'), 'company_posts', ['title'], unique=False)
+    op.create_index(op.f('ix_company_posts_work_experience'), 'company_posts', ['work_experience'], unique=False)
+    op.create_index(op.f('ix_company_posts_work_location'), 'company_posts', ['work_location'], unique=False)
+    op.create_index(op.f('ix_company_posts_working_hours'), 'company_posts', ['working_hours'], unique=False)
+    op.create_table('company_profile',
+    sa.Column('company_id', sa.Integer(), nullable=False),
+    sa.Column('industry_type', sa.String(), nullable=False),
+    sa.Column('employee_count', sa.Integer(), nullable=False),
+    sa.Column('establishment_date', sa.Date(), nullable=False),
+    sa.Column('company_type', sa.String(), nullable=False),
+    sa.Column('insurance', sa.String(), nullable=False),
+    sa.Column('phone_number', sa.String(), nullable=False),
+    sa.Column('address', sa.String(), nullable=False),
+    sa.Column('website_url', sa.String(), nullable=False),
+    sa.Column('email', sa.String(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['company_id'], ['companies.id'], ),
+    sa.PrimaryKeyConstraint('company_id')
+    )
+    op.create_index(op.f('ix_company_profile_address'), 'company_profile', ['address'], unique=False)
+    op.create_index(op.f('ix_company_profile_company_id'), 'company_profile', ['company_id'], unique=False)
+    op.create_index(op.f('ix_company_profile_company_type'), 'company_profile', ['company_type'], unique=False)
+    op.create_index(op.f('ix_company_profile_email'), 'company_profile', ['email'], unique=False)
+    op.create_index(op.f('ix_company_profile_employee_count'), 'company_profile', ['employee_count'], unique=False)
+    op.create_index(op.f('ix_company_profile_establishment_date'), 'company_profile', ['establishment_date'], unique=False)
+    op.create_index(op.f('ix_company_profile_industry_type'), 'company_profile', ['industry_type'], unique=False)
+    op.create_index(op.f('ix_company_profile_insurance'), 'company_profile', ['insurance'], unique=False)
+    op.create_index(op.f('ix_company_profile_phone_number'), 'company_profile', ['phone_number'], unique=False)
+    op.create_index(op.f('ix_company_profile_website_url'), 'company_profile', ['website_url'], unique=False)
+    op.create_table('company_users',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('company_id', sa.Integer(), nullable=False),
+    sa.Column('email', sa.String(), nullable=False),
+    sa.Column('password', sa.String(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('phone', sa.String(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['company_id'], ['companies.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_company_users_company_id'), 'company_users', ['company_id'], unique=False)
+    op.create_index(op.f('ix_company_users_email'), 'company_users', ['email'], unique=True)
+    op.create_index(op.f('ix_company_users_id'), 'company_users', ['id'], unique=False)
+    op.create_index(op.f('ix_company_users_name'), 'company_users', ['name'], unique=False)
+    op.create_index(op.f('ix_company_users_password'), 'company_users', ['password'], unique=False)
+    op.create_index(op.f('ix_company_users_phone'), 'company_users', ['phone'], unique=False)
+    op.create_table('profiles',
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('profile_image_url', sa.String(), nullable=True),
+    sa.Column('location', sa.String(), nullable=True),
+    sa.Column('introduction', sa.String(), nullable=True),
+    sa.Column('address', sa.String(), nullable=True),
+    sa.Column('position_id', sa.Integer(), nullable=True),
+    sa.Column('job_status', sa.String(), nullable=True),
+    sa.Column('portfolio_url', sa.String(), nullable=True),
+    sa.Column('birth_date', sa.Date(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('country_id', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['country_id'], ['countries.id'], ),
+    sa.ForeignKeyConstraint(['position_id'], ['positions.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('user_id')
+    )
+    op.create_index(op.f('ix_profiles_address'), 'profiles', ['address'], unique=False)
+    op.create_index(op.f('ix_profiles_birth_date'), 'profiles', ['birth_date'], unique=False)
+    op.create_index(op.f('ix_profiles_country_id'), 'profiles', ['country_id'], unique=False)
+    op.create_index(op.f('ix_profiles_introduction'), 'profiles', ['introduction'], unique=False)
+    op.create_index(op.f('ix_profiles_job_status'), 'profiles', ['job_status'], unique=False)
+    op.create_index(op.f('ix_profiles_location'), 'profiles', ['location'], unique=False)
+    op.create_index(op.f('ix_profiles_name'), 'profiles', ['name'], unique=False)
+    op.create_index(op.f('ix_profiles_portfolio_url'), 'profiles', ['portfolio_url'], unique=False)
+    op.create_index(op.f('ix_profiles_position_id'), 'profiles', ['position_id'], unique=False)
+    op.create_index(op.f('ix_profiles_profile_image_url'), 'profiles', ['profile_image_url'], unique=False)
+    op.create_index(op.f('ix_profiles_user_id'), 'profiles', ['user_id'], unique=False)
+    op.create_table('account_config',
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('sns_message_notice', sa.Boolean(), nullable=False),
+    sa.Column('email_notice', sa.Boolean(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['profiles.user_id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('user_id')
+    )
+    op.create_index(op.f('ix_account_config_email_notice'), 'account_config', ['email_notice'], unique=False)
+    op.create_index(op.f('ix_account_config_sns_message_notice'), 'account_config', ['sns_message_notice'], unique=False)
+    op.create_index(op.f('ix_account_config_user_id'), 'account_config', ['user_id'], unique=False)
+    op.create_table('contacts',
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('phone_number', sa.String(), nullable=True),
+    sa.Column('github_url', sa.String(), nullable=True),
+    sa.Column('linkedin_url', sa.String(), nullable=True),
+    sa.Column('website_url', sa.String(), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['profiles.user_id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('user_id')
+    )
+    op.create_index(op.f('ix_contacts_github_url'), 'contacts', ['github_url'], unique=False)
+    op.create_index(op.f('ix_contacts_linkedin_url'), 'contacts', ['linkedin_url'], unique=False)
+    op.create_index(op.f('ix_contacts_phone_number'), 'contacts', ['phone_number'], unique=False)
+    op.create_index(op.f('ix_contacts_user_id'), 'contacts', ['user_id'], unique=False)
+    op.create_index(op.f('ix_contacts_website_url'), 'contacts', ['website_url'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_contacts_website_url'), table_name='contacts')
+    op.drop_index(op.f('ix_contacts_user_id'), table_name='contacts')
+    op.drop_index(op.f('ix_contacts_phone_number'), table_name='contacts')
+    op.drop_index(op.f('ix_contacts_linkedin_url'), table_name='contacts')
+    op.drop_index(op.f('ix_contacts_github_url'), table_name='contacts')
+    op.drop_table('contacts')
+    op.drop_index(op.f('ix_account_config_user_id'), table_name='account_config')
+    op.drop_index(op.f('ix_account_config_sns_message_notice'), table_name='account_config')
+    op.drop_index(op.f('ix_account_config_email_notice'), table_name='account_config')
+    op.drop_table('account_config')
+    op.drop_index(op.f('ix_profiles_user_id'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_profile_image_url'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_position_id'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_portfolio_url'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_name'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_location'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_job_status'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_introduction'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_country_id'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_birth_date'), table_name='profiles')
+    op.drop_index(op.f('ix_profiles_address'), table_name='profiles')
+    op.drop_table('profiles')
+    op.drop_index(op.f('ix_company_users_phone'), table_name='company_users')
+    op.drop_index(op.f('ix_company_users_password'), table_name='company_users')
+    op.drop_index(op.f('ix_company_users_name'), table_name='company_users')
+    op.drop_index(op.f('ix_company_users_id'), table_name='company_users')
+    op.drop_index(op.f('ix_company_users_email'), table_name='company_users')
+    op.drop_index(op.f('ix_company_users_company_id'), table_name='company_users')
+    op.drop_table('company_users')
+    op.drop_index(op.f('ix_company_profile_website_url'), table_name='company_profile')
+    op.drop_index(op.f('ix_company_profile_phone_number'), table_name='company_profile')
+    op.drop_index(op.f('ix_company_profile_insurance'), table_name='company_profile')
+    op.drop_index(op.f('ix_company_profile_industry_type'), table_name='company_profile')
+    op.drop_index(op.f('ix_company_profile_establishment_date'), table_name='company_profile')
+    op.drop_index(op.f('ix_company_profile_employee_count'), table_name='company_profile')
+    op.drop_index(op.f('ix_company_profile_email'), table_name='company_profile')
+    op.drop_index(op.f('ix_company_profile_company_type'), table_name='company_profile')
+    op.drop_index(op.f('ix_company_profile_company_id'), table_name='company_profile')
+    op.drop_index(op.f('ix_company_profile_address'), table_name='company_profile')
+    op.drop_table('company_profile')
+    op.drop_index(op.f('ix_company_posts_working_hours'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_work_location'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_work_experience'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_title'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_start_date'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_salary'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_position_id'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_language'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_id'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_end_date'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_employment_type'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_education'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_content'), table_name='company_posts')
+    op.drop_index(op.f('ix_company_posts_company_id'), table_name='company_posts')
+    op.drop_table('company_posts')
+    op.drop_index(op.f('ix_users_passport_certi'), table_name='users')
+    op.drop_index(op.f('ix_users_id'), table_name='users')
+    op.drop_index(op.f('ix_users_email'), table_name='users')
+    op.drop_table('users')
+    op.drop_index(op.f('ix_positions_parent_id'), table_name='positions')
+    op.drop_index(op.f('ix_positions_name'), table_name='positions')
+    op.drop_index(op.f('ix_positions_level'), table_name='positions')
+    op.drop_index(op.f('ix_positions_id'), table_name='positions')
+    op.drop_index(op.f('ix_positions_code'), table_name='positions')
+    op.drop_table('positions')
+    op.drop_index(op.f('ix_countries_name'), table_name='countries')
+    op.drop_index(op.f('ix_countries_id'), table_name='countries')
+    op.drop_index(op.f('ix_countries_code'), table_name='countries')
+    op.drop_table('countries')
+    op.drop_index(op.f('ix_companies_id'), table_name='companies')
+    op.drop_index(op.f('ix_companies_company_number'), table_name='companies')
+    op.drop_index(op.f('ix_companies_company_name'), table_name='companies')
+    op.drop_table('companies')
+    # ### end Alembic commands ###
