--- alembic/versions/4caeeb791f47_fix_update_the_company_number_column_.py
@@ -1,46 +1,0 @@
-"""fix:Update the company_number column and phone from integer to large integer 
-
-Revision ID: 4caeeb791f47
-Revises: b41ed4b86b5c
-Create Date: 2025-10-30 19:34:46.962580
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '4caeeb791f47'
-down_revision: Union[str, Sequence[str], None] = 'b41ed4b86b5c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('companies', 'company_number',
-               existing_type=sa.INTEGER(),
-               type_=sa.BigInteger(),
-               existing_nullable=False)
-    op.alter_column('company_users', 'phone',
-               existing_type=sa.INTEGER(),
-               type_=sa.BigInteger(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('company_users', 'phone',
-               existing_type=sa.BigInteger(),
-               type_=sa.INTEGER(),
-               existing_nullable=False)
-    op.alter_column('companies', 'company_number',
-               existing_type=sa.BigInteger(),
-               type_=sa.INTEGER(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
