--- alembic/versions/b41ed4b86b5c_fix_update_company_table_add_company_.py
@@ -1,78 +1,0 @@
-"""fix:update company table, Add company_user table
-
-Revision ID: b41ed4b86b5c
-Revises: 05cfcb09c1ff
-Create Date: 2025-10-29 02:19:51.410651
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'b41ed4b86b5c'
-down_revision: Union[str, Sequence[str], None] = '05cfcb09c1ff'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('company_users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('company_id', sa.Integer(), nullable=False),
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('password', sa.String(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('phone', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['company_id'], ['companies.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_company_users_company_id'), 'company_users', ['company_id'], unique=False)
-    op.create_index(op.f('ix_company_users_email'), 'company_users', ['email'], unique=True)
-    op.create_index(op.f('ix_company_users_id'), 'company_users', ['id'], unique=False)
-    op.create_index(op.f('ix_company_users_name'), 'company_users', ['name'], unique=False)
-    op.create_index(op.f('ix_company_users_password'), 'company_users', ['password'], unique=False)
-    op.create_index(op.f('ix_company_users_phone'), 'company_users', ['phone'], unique=False)
-    op.drop_index(op.f('ix_refresh_tokens_expires_at'), table_name='refresh_tokens')
-    op.drop_index(op.f('ix_refresh_tokens_id'), table_name='refresh_tokens')
-    op.drop_index(op.f('ix_refresh_tokens_token'), table_name='refresh_tokens')
-    op.drop_index(op.f('ix_refresh_tokens_user_id'), table_name='refresh_tokens')
-    op.drop_table('refresh_tokens')
-    op.drop_column('companies', 'managers')
-    op.drop_column('users', 'company_info')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('company_info', postgresql.JSONB(astext_type=sa.Text()), autoincrement=False, nullable=True))
-    op.add_column('companies', sa.Column('managers', postgresql.JSONB(astext_type=sa.Text()), autoincrement=False, nullable=False))
-    op.create_table('refresh_tokens',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('token', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('user_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('expires_at', postgresql.TIMESTAMP(), autoincrement=False, nullable=False),
-    sa.Column('created_at', postgresql.TIMESTAMP(timezone=True), autoincrement=False, nullable=False),
-    sa.Column('updated_at', postgresql.TIMESTAMP(timezone=True), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], name=op.f('refresh_tokens_user_id_fkey')),
-    sa.PrimaryKeyConstraint('id', name=op.f('refresh_tokens_pkey'))
-    )
-    op.create_index(op.f('ix_refresh_tokens_user_id'), 'refresh_tokens', ['user_id'], unique=False)
-    op.create_index(op.f('ix_refresh_tokens_token'), 'refresh_tokens', ['token'], unique=True)
-    op.create_index(op.f('ix_refresh_tokens_id'), 'refresh_tokens', ['id'], unique=False)
-    op.create_index(op.f('ix_refresh_tokens_expires_at'), 'refresh_tokens', ['expires_at'], unique=False)
-    op.drop_index(op.f('ix_company_users_phone'), table_name='company_users')
-    op.drop_index(op.f('ix_company_users_password'), table_name='company_users')
-    op.drop_index(op.f('ix_company_users_name'), table_name='company_users')
-    op.drop_index(op.f('ix_company_users_id'), table_name='company_users')
-    op.drop_index(op.f('ix_company_users_email'), table_name='company_users')
-    op.drop_index(op.f('ix_company_users_company_id'), table_name='company_users')
-    op.drop_table('company_users')
-    # ### end Alembic commands ###
